pyproject.toml
[project]
name = "ml-engineer"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.13.5",
    "chromadb>=1.0.20",
    "google-api-python-client>=2.181.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "html2text>=2025.4.15",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "scikit-learn>=1.3.0",
    "sentence-transformers>=5.1.0",
    "streamlit>=1.49.1",
    "torch>=2.5.1",
    "transformers[torch]>=4.44.0",
]


[tool.uv.sources]
torch = [
    { index = "pytorch-gpu" },
]

[[tool.uv.index]]
name = "pytorch-gpu"
url = "https://download.pytorch.org/whl/cu121"
explicit = true
